update angular to latest
