+++ angular-boilerplate/src/app/app.module.ts
@@ -1,9 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
-
 import { AppComponent } from './app.component';
-
 
 @NgModule({
   declarations: [
