+++ angular-boilerplate/src/favicon.ico
